move s3 to source directory
