+++ packages/ragbits-document-search/src/ragbits/document_search/documents/sources/__init__.py
@@ -3,5 +3,6 @@ from ragbits.document_search.documents.sources.azure import AzureBlobStorageSour
 from ragbits.document_search.documents.sources.gcs import GCSSource
 from ragbits.document_search.documents.sources.hf import HuggingFaceSource
 from ragbits.document_search.documents.sources.local import LocalFileSource
+from ragbits.document_search.documents.sources.s3 import S3Source
 
-__all__ = ["AzureBlobStorageSource", "GCSSource", "HuggingFaceSource", "LocalFileSource", "Source"]
+__all__ = ["AzureBlobStorageSource", "GCSSource", "HuggingFaceSource", "LocalFileSource", "S3Source", "Source"]
